+++ ts/components/game/HeroPower.tsx
@@ -3,6 +3,7 @@ import * as React from "react";
 import {EntityInPlayProps} from "../../interfaces";
 import EntityInPlay from "./EntityInPlay";
 import Cost from "./stats/Cost";
+import Card from "./Card"
 import HeroPowerArt from "./visuals/HeroPowerArt";
 import {GameTag} from "../../enums";
 
@@ -36,6 +37,20 @@ class HeroPower extends EntityInPlay<EntityInPlayProps> {
 				cardArtDirectory={this.props.cardArtDirectory}
 			/>
 		];
+
+		if (this.state.isHovering) {
+			components.push(<div key="hover" className="mouse-over">
+				<Card
+					entity={entity}
+					assetDirectory={this.props.assetDirectory}
+					cards={this.props.cards}
+					isHidden={false}
+					controller={this.props.controller}
+					cardArtDirectory={this.props.cardArtDirectory}
+					option={null}
+					defaultStats={true}
+				/></div>);
+		}
 
 		if (!entity.isExhausted()) {
 			components.push(<Cost